lambda/solarEdgeGeneric/requirements.txt
psycopg2_binary == 2.8.6
pytz == 2021.1
requests == 2.22.0
